+++ lib/console-log-strategy/simple-console-log.test.d.ts
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=simple-console-log.test.d.ts.map