+++ lib/console-logger.d.ts
@@ -1,21 +1,20 @@
-import { LoggerStrategy } from './logger-strategy';
-export declare enum LogLevelType {
-    ERROR = "error",
-    WARN = "warn",
-    INFO = "info",
-    DEBUG = "debug"
-}
-export declare type LogLevel = 'error' | 'warn' | 'info' | 'debug' | 'ERROR' | 'WARN' | 'INFO' | 'DEBUG';
+import { ConsoleLogStrategy } from './console-log-strategy/console-log-strategy';
+import { LogLevelStringType, LogLevelType } from './log-level-type';
+import { LoggerStrategy, ObjectType, StringOrObjectType } from './logger-strategy';
+export declare type ConsoleLoggerParams = {
+    logLevel?: LogLevelType | LogLevelStringType;
+    consoleLogStrategy?: ConsoleLogStrategy;
+};
 export declare class ConsoleLogger implements LoggerStrategy {
-    protected _logLevel: LogLevelType;
-    constructor(logLevel?: LogLevelType | LogLevel);
+    protected readonly _logLevel: LogLevelType;
+    protected readonly _consoleLogStrategy: ConsoleLogStrategy;
+    constructor({ logLevel, consoleLogStrategy }?: ConsoleLoggerParams);
     protected _logLevelToInt: (logLevel: LogLevelType) => number;
     protected _shouldLog: (currentLevel: LogLevelType) => boolean;
-    protected _consoleLog(msg: any, obj?: any): void;
-    protected _logMessage: (type: LogLevelType, msg: any, obj?: any) => void;
-    debug(msg: any, obj?: any): void;
-    info(msg: any, obj?: any): void;
-    warn(msg: any, obj?: any): void;
-    error(msg: any, obj?: any): void;
+    protected _logMessage: (type: LogLevelType, messageObject: StringOrObjectType, meta?: ObjectType | undefined) => void;
+    debug(messageObject: StringOrObjectType, meta?: ObjectType): void;
+    info(messageObject: StringOrObjectType, meta?: ObjectType): void;
+    warn(messageObject: StringOrObjectType, meta?: ObjectType): void;
+    error(messageObject: StringOrObjectType, meta?: ObjectType): void;
 }
 //# sourceMappingURL=console-logger.d.ts.map