+++ lib/index.d.ts
@@ -1,4 +1,3 @@
-export * from './console-logger';
 export * from './logger-strategy';
-export * from './no-logger';
+export * from './log-level-type';
 //# sourceMappingURL=index.d.ts.map