+++ lib/log-level-type.d.ts
@@ -1,0 +1,8 @@
+export declare enum LogLevelType {
+    ERROR = "error",
+    WARN = "warn",
+    INFO = "info",
+    DEBUG = "debug"
+}
+export declare type LogLevelStringType = 'error' | 'warn' | 'info' | 'debug' | 'ERROR' | 'WARN' | 'INFO' | 'DEBUG';
+//# sourceMappingURL=log-level-type.d.ts.map