+++ lib/no-logger.d.ts
@@ -1,8 +1,8 @@
-import { LoggerStrategy } from './logger-strategy';
+import { LoggerStrategy, ObjectType, StringOrObjectType } from './logger-strategy';
 export declare class NoLogger implements LoggerStrategy {
-    debug(_: any, __?: any): void;
-    error(_: any, __?: any): void;
-    info(_: any, __?: any): void;
-    warn(_: any, __?: any): void;
+    debug(_: StringOrObjectType, __?: ObjectType): void;
+    error(_: StringOrObjectType, __?: ObjectType): void;
+    info(_: StringOrObjectType, __?: ObjectType): void;
+    warn(_: StringOrObjectType, __?: ObjectType): void;
 }
 //# sourceMappingURL=no-logger.d.ts.map